Extract `check-modifiers`
