+++ packages/typescript-estree/src/convert.ts
@@ -10,13 +10,13 @@ import type {
 import type { SemanticOrSyntacticError } from './semantic-or-syntactic-errors';
 import type { TSESTree, TSESTreeToTSNode, TSNode } from './ts-estree';
 
+import { checkModifiers } from './check-modifiers';
 import { getDecorators, getModifiers } from './getModifiers';
 import {
   canContainDirective,
   createError,
   findNextToken,
   getBinaryExpressionType,
-  getContainingFunction,
   getDeclarationKind,
   getLastModifier,
   getLineAndCharacterFor,
@@ -34,9 +34,6 @@ import {
   isOptional,
   isThisInTypeQuery,
   isValidAssignmentTarget,
-  nodeCanBeDecorated,
-  nodeHasIllegalDecorators,
-  nodeIsPresent,
   unescapeStringLiteralText,
 } from './node-utils';
 import { AST_NODE_TYPES } from './ts-estree';
@@ -158,247 +155,7 @@ export class Converter {
       return;
     }
 
-    // typescript<5.0.0
-    if (nodeHasIllegalDecorators(node)) {
-      this.#throwError(
-        node.illegalDecorators[0],
-        'Decorators are not valid here.',
-      );
-    }
-
-    for (const decorator of getDecorators(
-      node,
-      /* includeIllegalDecorators */ true,
-    ) ?? []) {
-      // `checkGrammarModifiers` function in typescript
-      if (!nodeCanBeDecorated(node as TSNode)) {
-        if (ts.isMethodDeclaration(node) && !nodeIsPresent(node.body)) {
-          this.#throwError(
-            decorator,
-            'A decorator can only decorate a method implementation, not an overload.',
-          );
-        } else {
-          this.#throwError(decorator, 'Decorators are not valid here.');
-        }
-      }
-    }
-
-    for (const modifier of getModifiers(
-      node,
-      /* includeIllegalModifiers */ true,
-    ) ?? []) {
-      if (modifier.kind !== SyntaxKind.ReadonlyKeyword) {
-        if (
-          node.kind === SyntaxKind.PropertySignature ||
-          node.kind === SyntaxKind.MethodSignature
-        ) {
-          this.#throwError(
-            modifier,
-            `'${ts.tokenToString(
-              modifier.kind,
-            )}' modifier cannot appear on a type member`,
-          );
-        }
-
-        if (
-          node.kind === SyntaxKind.IndexSignature &&
-          (modifier.kind !== SyntaxKind.StaticKeyword ||
-            !ts.isClassLike(node.parent))
-        ) {
-          this.#throwError(
-            modifier,
-            `'${ts.tokenToString(
-              modifier.kind,
-            )}' modifier cannot appear on an index signature`,
-          );
-        }
-      }
-
-      if (
-        modifier.kind !== SyntaxKind.InKeyword &&
-        modifier.kind !== SyntaxKind.OutKeyword &&
-        modifier.kind !== SyntaxKind.ConstKeyword &&
-        node.kind === SyntaxKind.TypeParameter
-      ) {
-        this.#throwError(
-          modifier,
-          `'${ts.tokenToString(
-            modifier.kind,
-          )}' modifier cannot appear on a type parameter`,
-        );
-      }
-
-      if (
-        (modifier.kind === SyntaxKind.InKeyword ||
-          modifier.kind === SyntaxKind.OutKeyword) &&
-        (node.kind !== SyntaxKind.TypeParameter ||
-          !(
-            ts.isInterfaceDeclaration(node.parent) ||
-            ts.isClassLike(node.parent) ||
-            ts.isTypeAliasDeclaration(node.parent)
-          ))
-      ) {
-        this.#throwError(
-          modifier,
-          `'${ts.tokenToString(
-            modifier.kind,
-          )}' modifier can only appear on a type parameter of a class, interface or type alias`,
-        );
-      }
-
-      if (
-        modifier.kind === SyntaxKind.ReadonlyKeyword &&
-        node.kind !== SyntaxKind.PropertyDeclaration &&
-        node.kind !== SyntaxKind.PropertySignature &&
-        node.kind !== SyntaxKind.IndexSignature &&
-        node.kind !== SyntaxKind.Parameter
-      ) {
-        this.#throwError(
-          modifier,
-          "'readonly' modifier can only appear on a property declaration or index signature.",
-        );
-      }
-
-      if (
-        modifier.kind === SyntaxKind.DeclareKeyword &&
-        ts.isClassLike(node.parent) &&
-        !ts.isPropertyDeclaration(node)
-      ) {
-        this.#throwError(
-          modifier,
-          `'${ts.tokenToString(
-            modifier.kind,
-          )}' modifier cannot appear on class elements of this kind.`,
-        );
-      }
-
-      if (
-        modifier.kind === SyntaxKind.DeclareKeyword &&
-        ts.isVariableStatement(node)
-      ) {
-        const declarationKind = getDeclarationKind(node.declarationList);
-        if (declarationKind === 'using' || declarationKind === 'await using') {
-          this.#throwError(
-            modifier,
-            `'declare' modifier cannot appear on a '${declarationKind}' declaration.`,
-          );
-        }
-      }
-
-      if (
-        modifier.kind === SyntaxKind.AbstractKeyword &&
-        node.kind !== SyntaxKind.ClassDeclaration &&
-        node.kind !== SyntaxKind.ConstructorType &&
-        node.kind !== SyntaxKind.MethodDeclaration &&
-        node.kind !== SyntaxKind.PropertyDeclaration &&
-        node.kind !== SyntaxKind.GetAccessor &&
-        node.kind !== SyntaxKind.SetAccessor
-      ) {
-        this.#throwError(
-          modifier,
-          `'${ts.tokenToString(
-            modifier.kind,
-          )}' modifier can only appear on a class, method, or property declaration.`,
-        );
-      }
-
-      if (
-        (modifier.kind === SyntaxKind.StaticKeyword ||
-          modifier.kind === SyntaxKind.PublicKeyword ||
-          modifier.kind === SyntaxKind.ProtectedKeyword ||
-          modifier.kind === SyntaxKind.PrivateKeyword) &&
-        (node.parent.kind === SyntaxKind.ModuleBlock ||
-          node.parent.kind === SyntaxKind.SourceFile)
-      ) {
-        this.#throwError(
-          modifier,
-          `'${ts.tokenToString(
-            modifier.kind,
-          )}' modifier cannot appear on a module or namespace element.`,
-        );
-      }
-
-      if (
-        modifier.kind === SyntaxKind.AccessorKeyword &&
-        node.kind !== SyntaxKind.PropertyDeclaration
-      ) {
-        this.#throwError(
-          modifier,
-          "'accessor' modifier can only appear on a property declaration.",
-        );
-      }
-
-      // `checkGrammarAsyncModifier` function in `typescript`
-      if (
-        modifier.kind === SyntaxKind.AsyncKeyword &&
-        node.kind !== SyntaxKind.MethodDeclaration &&
-        node.kind !== SyntaxKind.FunctionDeclaration &&
-        node.kind !== SyntaxKind.FunctionExpression &&
-        node.kind !== SyntaxKind.ArrowFunction
-      ) {
-        this.#throwError(modifier, "'async' modifier cannot be used here.");
-      }
-
-      // `checkGrammarModifiers` function in `typescript`
-      if (
-        node.kind === SyntaxKind.Parameter &&
-        (modifier.kind === SyntaxKind.StaticKeyword ||
-          modifier.kind === SyntaxKind.ExportKeyword ||
-          modifier.kind === SyntaxKind.DeclareKeyword ||
-          modifier.kind === SyntaxKind.AsyncKeyword)
-      ) {
-        this.#throwError(
-          modifier,
-          `'${ts.tokenToString(
-            modifier.kind,
-          )}' modifier cannot appear on a parameter.`,
-        );
-      }
-
-      // `checkGrammarModifiers` function in `typescript`
-      if (
-        modifier.kind === SyntaxKind.PublicKeyword ||
-        modifier.kind === SyntaxKind.ProtectedKeyword ||
-        modifier.kind === SyntaxKind.PrivateKeyword
-      ) {
-        for (const anotherModifier of getModifiers(node) ?? []) {
-          if (
-            anotherModifier !== modifier &&
-            (anotherModifier.kind === SyntaxKind.PublicKeyword ||
-              anotherModifier.kind === SyntaxKind.ProtectedKeyword ||
-              anotherModifier.kind === SyntaxKind.PrivateKeyword)
-          ) {
-            this.#throwError(
-              anotherModifier,
-              `Accessibility modifier already seen.`,
-            );
-          }
-        }
-      }
-
-      // `checkParameter` function in `typescript`
-      if (
-        node.kind === SyntaxKind.Parameter &&
-        // In `typescript` package, it's `ts.hasSyntacticModifier(node, ts.ModifierFlags.ParameterPropertyModifier)`
-        // https://github.com/typescript-eslint/typescript-eslint/pull/6615#discussion_r1136489935
-        (modifier.kind === SyntaxKind.PublicKeyword ||
-          modifier.kind === SyntaxKind.PrivateKeyword ||
-          modifier.kind === SyntaxKind.ProtectedKeyword ||
-          modifier.kind === SyntaxKind.ReadonlyKeyword ||
-          modifier.kind === SyntaxKind.OverrideKeyword)
-      ) {
-        const func = getContainingFunction(node)!;
-
-        if (
-          !(func.kind === SyntaxKind.Constructor && nodeIsPresent(func.body))
-        ) {
-          this.#throwError(
-            modifier,
-            'A parameter property is only allowed in a constructor implementation.',
-          );
-        }
-      }
-    }
+    checkModifiers(node);
   }
 
   #throwError(node: number | ts.Node | TSESTree.Range, message: string): never {
